Sorted commands into folders & misc - Sorted commands into folder corresponding to their viewModel - Reformatted some xaml files - Variable naming fix
