+++ BackupUtil.ViewModel/Command/CreateJob/CreateJobCommand.cs
@@ -4,7 +4,7 @@ using BackupUtil.ViewModel.Service;
 using BackupUtil.ViewModel.Store;
 using BackupUtil.ViewModel.ViewModel;
 
-namespace BackupUtil.ViewModel.Command;
+namespace BackupUtil.ViewModel.Command.CreateJob;
 
 public class CreateJobCommand<TViewModel> : CommandBase where TViewModel : ViewModelBase
 {
