+++ BackupUtil.ViewModel/Command/JobListing/DeleteSelectedJobsCommand.cs
@@ -1,29 +1,29 @@
-﻿using System.ComponentModel;
+using System.ComponentModel;
 using BackupUtil.ViewModel.Store;
 using BackupUtil.ViewModel.ViewModel;
 
-namespace BackupUtil.ViewModel.Command;
+namespace BackupUtil.ViewModel.Command.JobListing;
 
-public class CreateTransactionsForSelectedJobsCommand : CommandBase
+public class DeleteSelectedJobsCommand : CommandBase
 {
     private readonly JobListingViewModel _jobListingViewModel;
     private readonly JobStore _jobStore;
 
-    public CreateTransactionsForSelectedJobsCommand(JobListingViewModel jobListingViewModel, JobStore jobStore)
+    public DeleteSelectedJobsCommand(JobListingViewModel jobListingViewModel, JobStore jobStore)
     {
         _jobStore = jobStore;
         _jobListingViewModel = jobListingViewModel;
         _jobListingViewModel.PropertyChanged += OnViewModelPropertyChanged;
     }
 
+    public override void Execute(object? parameter)
+    {
+        _jobStore.RemoveByIndices(_jobListingViewModel.SelectedJobIndices);
+    }
+
     public override bool CanExecute(object? parameter)
     {
         return _jobListingViewModel.SelectedJobIndices.Count > 0 && base.CanExecute(parameter);
-    }
-
-    public override void Execute(object? parameter)
-    {
-        _jobStore.RunByIndices(_jobListingViewModel.SelectedJobIndices);
     }
 
     private void OnViewModelPropertyChanged(object? sender, PropertyChangedEventArgs e)
