+++ BackupUtil.ViewModel/Command/JobListing/LoadJobsCommand.cs
@@ -1,0 +1,38 @@
+﻿using System.ComponentModel;
+using BackupUtil.ViewModel.Store;
+using BackupUtil.ViewModel.ViewModel;
+
+namespace BackupUtil.ViewModel.Command.JobListing;
+
+public class LoadJobsCommand : CommandBase
+{
+    private readonly JobListingViewModel _jobListingViewModel;
+    private readonly JobStore _jobManager;
+
+    public LoadJobsCommand(JobListingViewModel jobListingViewModel, JobStore jobManager)
+    {
+        _jobManager = jobManager;
+        _jobListingViewModel = jobListingViewModel;
+        _jobListingViewModel.PropertyChanged += OnViewModelPropertyChanged;
+    }
+
+    public override bool CanExecute(object? parameter)
+    {
+        return _jobListingViewModel.CanAccessJobFile && base.CanExecute(parameter);
+    }
+
+    public override void Execute(object? parameter)
+    {
+        _jobManager.RemoveAll();
+        _jobManager.LoadJobs();
+        _jobListingViewModel.LoadJobViewModels();
+    }
+
+    private void OnViewModelPropertyChanged(object? sender, PropertyChangedEventArgs e)
+    {
+        if (e.PropertyName == nameof(_jobListingViewModel.CanAccessJobFile))
+        {
+            OnCanExecuteChanged();
+        }
+    }
+}
