+++ BackupUtil.ViewModel/Command/Settings/ApplySettingsCommand.cs
@@ -3,7 +3,7 @@ using BackupUtil.ViewModel.Service;
 using BackupUtil.ViewModel.Store;
 using BackupUtil.ViewModel.ViewModel;
 
-namespace BackupUtil.ViewModel.Command;
+namespace BackupUtil.ViewModel.Command.Settings;
 
 public class ApplySettingsCommand<TViewModel> : CommandBase where TViewModel : ViewModelBase
 {
